Moving type into module
