+++ src/services/logging/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import type { Logger } from 'hyperview/src/types';
+import type { Logger } from './types';
 
 let logger: Logger = console;
 
@@ -13,3 +13,5 @@ export const log = (m?: any, ...p: any[]): void => logger.log(m, p);
 export const info = (m?: any, ...p: any[]): void => logger.info(m, p);
 export const warn = (m?: any, ...p: any[]): void => logger.warn(m, p);
 export const error = (m?: any, ...p: any[]): void => logger.error(m, p);
+
+export type { Logger } from './types';
