+++ src/types.ts
@@ -431,13 +431,3 @@ export type Reload = (
   optHref: DOMString | null | undefined,
   opts: HvComponentOptions,
 ) => void;
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-type LogFunction = (message?: any, ...optionalParams: any[]) => void;
-
-export type Logger = {
-  error: LogFunction;
-  info: LogFunction;
-  log: LogFunction;
-  warn: LogFunction;
-};
